backend/requirements.txt
Django>=4.2
djangorestframework
djangorestframework-simplejwt
psycopg2-binary
django-cors-headers
python-dotenv
gunicorn
